alpine-mingw

+++ pg_dump/build_mingw.sh
@@ -17,9 +17,16 @@ echo "::set-output name=tool_version::$tool_version"
 download_url="https://github.com/postgres/postgres/archive/refs/tags/$tool_version.tar.gz"
 echo "::group::prepare sources $download_url"
 
-mkdir -p "$dp0/release/build" && cd "$dp0/release"
-wget "$download_url" -O "tool-$tool_version.tar.gz" && tar -xf "tool-$tool_version.tar.gz"
 tool_root_path="$dp0/release/postgres-$tool_version"
+
+mkdir -p "$dp0/release/build" && cd "$dp0/release"
+
+if [[ ! -f "$tool_root_path/configure" ]]; then
+  wget "$download_url" -O "tool-$tool_version.tar.gz" && tar -xf "tool-$tool_version.tar.gz"
+fi
+
+cp "$dp0/release/Makefile" "$tool_root_path/src/bin/pg_dump/"
+
 cd "$tool_root_path"
 
 echo "::endgroup::"
